+++ src/GlobalId/NodeRegistry.php
@@ -76,6 +76,14 @@ class NodeRegistry
     {
         [$decodedType, $decodedId] = $args['id'];
 
+        // This check forces Lighthouse to eagerly load the type, which might not have
+        // happened if the client only references it indirectly through an interface.
+        // Loading the type in turn causes the TypeMiddleware to run and thus register
+        // the type in the NodeRegistry.
+        if (! $this->typeRegistry->has($decodedType)) {
+            throw new Error("[{$decodedType}] is not a type and cannot be resolved.");
+        }
+
         // Check if we have a resolver registered for the given type
         if (! $resolver = Arr::get($this->nodeResolver, $decodedType)) {
             throw new Error("[{$decodedType}] is not a registered node and cannot be resolved.");