+++ src/Pagination/PaginationType.php
@@ -9,11 +9,11 @@ use Nuwave\Lighthouse\Exceptions\DefinitionException;
  */
 class PaginationType
 {
-    public const TYPE_PAGINATOR = 'paginator';
-    public const PAGINATION_TYPE_CONNECTION = 'connection';
+    public const PAGINATOR = 'PAGINATOR';
+    public const CONNECTION = 'CONNECTION';
 
     /**
-     * @var string PAGINATION_TYPE_PAGINATOR|PAGINATION_TYPE_CONNECTION
+     * @var string One of the constant values in this class
      */
     protected $type;
 
@@ -22,14 +22,15 @@ class PaginationType
      */
     public function __construct(string $paginationType)
     {
-        switch ($paginationType) {
+        // TODO remove lowercase and alternate options in v6
+        switch (strtolower($paginationType)) {
             case 'default':
             case 'paginator':
-                $this->type = self::TYPE_PAGINATOR;
+                $this->type = self::PAGINATOR;
                 break;
             case 'connection':
             case 'relay':
-                $this->type = self::PAGINATION_TYPE_CONNECTION;
+                $this->type = self::CONNECTION;
                 break;
             default:
                 throw new DefinitionException(
@@ -40,11 +41,11 @@ class PaginationType
 
     public function isPaginator(): bool
     {
-        return $this->type === self::TYPE_PAGINATOR;
+        return $this->type === self::PAGINATOR;
     }
 
     public function isConnection(): bool
     {
-        return $this->type === self::PAGINATION_TYPE_CONNECTION;
+        return $this->type === self::CONNECTION;
     }
 }