+++ src/Schema/Directives/CountDirective.php
@@ -2,12 +2,14 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
-use Illuminate\Database\Eloquent\Builder;
+use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Database\Eloquent\Model;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
-use Nuwave\Lighthouse\Execution\DataLoader\RelationCountBatchLoader;
+use Nuwave\Lighthouse\Execution\DataLoader\RelationCountLoader;
+use Nuwave\Lighthouse\Execution\DataLoader\RelationLoader;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
+use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 class CountDirective extends WithRelationDirective implements FieldResolver
 {
@@ -20,11 +22,13 @@ Returns the count of a given relationship or model.
 directive @count(
   """
   The relationship which you want to run the count on.
+  Mutually exclusive with the `model` argument.
   """
   relation: String
 
   """
   The model to run the count on.
+  Mutually exclusive with the `relation` argument.
   """
   model: String
 
@@ -38,50 +42,53 @@ GRAPHQL;
 
     public function resolveField(FieldValue $value): FieldValue
     {
-        return $value->setResolver(
-            $this->deferredRelationResolver(
-                function (?Model $model) {
-                    // Fetch the count by relation
-                    $relation = $this->directiveArgValue('relation');
-                    if (! is_null($relation)) {
-                        return $model->{$this->nodeName()};
-                    }
+        $modelArg = $this->directiveArgValue('model');
+        if (! is_null($modelArg)) {
+            return $value->setResolver(
+                function ($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($modelArg): int {
+                    /** @var \Illuminate\Database\Eloquent\Builder $query */
+                    $query = $this
+                        ->namespaceModelClass($modelArg)
+                        ::query();
 
-                    // Else we try to fetch by model.
-                    $modelArg = $this->directiveArgValue('model');
-                    if (! is_null($modelArg)) {
-                        return $this->countModel($modelArg);
-                    }
+                    $this->decorateBuilder($resolveInfo)($query);
 
-                    throw new DefinitionException(
-                        "A `model` or `relation` argument must be assigned to the '{$this->name()}' directive on '{$this->nodeName()}"
-                    );
+                    return $query->count();
                 }
-            )
+            );
+        }
+
+        // Fetch the count by relation
+        $relation = $this->directiveArgValue('relation');
+        if (! is_null($relation)) {
+            return $value->setResolver(
+                function (Model $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) {
+                    return $this->loadRelation($parent, $resolveInfo);
+                }
+            );
+        }
+
+        throw new DefinitionException(
+            "A `model` or `relation` argument must be assigned to the '{$this->name()}' directive on '{$this->nodeName()}'."
         );
     }
 
-    public function batchLoaderClass(): string
+    protected function relationName(): string
     {
-        return RelationCountBatchLoader::class;
-    }
-
-    protected function countModel(string $modelName): int
-    {
-        $scopesArg = $this->directiveArgValue('scopes');
-
-        return $this->namespaceModelClass($modelName)
-            ::query()
-            ->when(! is_null($scopesArg), function (Builder $query) use ($scopesArg) {
-                return $query->scopes($scopesArg);
-            })
-            ->count();
-    }
-
-    public function relationName(): string
-    {
+        /**
+         * We only got to this point because we already know this argument is set.
+         *
+         * @var string $relation
+         */
         $relation = $this->directiveArgValue('relation');
 
-        return "{$relation} as {$this->nodeName()}";
+        return $relation;
+    }
+
+    protected function relationLoader(ResolveInfo $resolveInfo): RelationLoader
+    {
+        return new RelationCountLoader(
+            $this->decorateBuilder($resolveInfo)
+        );
     }
 }