+++ src/Schema/Directives/MutationExecutorDirective.php
@@ -11,30 +11,19 @@ use Nuwave\Lighthouse\Execution\Arguments\ResolveNested;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\ArgResolver;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
-use Nuwave\Lighthouse\Support\Contracts\GlobalId;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Nuwave\Lighthouse\Support\Utils;
 
 abstract class MutationExecutorDirective extends BaseDirective implements FieldResolver, ArgResolver
 {
     /**
-     * The database manager.
-     *
      * @var \Illuminate\Database\DatabaseManager
      */
     protected $databaseManager;
 
-    /**
-     * The GlobalId resolver.
-     *
-     * @var \Nuwave\Lighthouse\Support\Contracts\GlobalId
-     */
-    protected $globalId;
-
-    public function __construct(DatabaseManager $databaseManager, GlobalId $globalId)
+    public function __construct(DatabaseManager $databaseManager)
     {
         $this->databaseManager = $databaseManager;
-        $this->globalId = $globalId;
     }
 
     public function resolveField(FieldValue $fieldValue): FieldValue