+++ tests/Integration/SchemaCachingTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Integration;
 
+use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Nuwave\Lighthouse\Schema\AST\ASTBuilder;
 use Tests\TestCase;
 use Tests\TestsSerialization;
@@ -16,7 +17,7 @@ class SchemaCachingTest extends TestCase
         parent::getEnvironmentSetUp($app);
 
         /** @var \Illuminate\Contracts\Config\Repository $config */
-        $config = $app['config'];
+        $config = $app->make(ConfigRepository::class);
         $config->set('lighthouse.cache.enable', true);
 
         $this->useSerializingArrayStore($app);